gwnet_http: Fix set HTTP code function
The parameter must be at least uint16_t as the HTTP code can
be more than 255.

Fixes: c3fd714c0525227a06c1e5b62a4a1be9fd3d295c ("Full project rewrite")

+++ gwnet_http.c
@@ -234,7 +234,7 @@ struct gwnet_http_res *gwnet_http_req_get_res(gwnet_http_req_t *req)
 	return &req->res;
 }
 
-void gwnet_http_res_set_code(gwnet_http_res_t *res, uint8_t code)
+void gwnet_http_res_set_code(gwnet_http_res_t *res, uint16_t code)
 {
 	res->status = code;
 }
